refactor: bookstore backend restart
This is the first commit of SE3353 E-Bookstore backend.
However, it's migrated from the backend project of SE2321.
The backend project is restarted.

Here lists what the commit does:

1. Migrate from maven to gradle.
Gradle is a modern build system, and it's also recommended by the Spring official.
Kotlin script is used to configure the build system.

2. Remove two unnecessary dependencies.
They are _commons-io_ and _org.modelmapper.modelmapper_.
Related codes are also fixed.

3. Boost the JDK and Spring version.
JDK: 11 -> 17
Spring Boot: 2.7 -> 3.1
Since Jakarta EE 9 is release, javax.* is moved to jakarta.*,
so the imports are modified.

4. Resolve the new problem after migration.
The problem is that jackson can't deserialize JSON to the object
that doesn't have a contructor with empty arguments.
The solution is prepending the annotation `@NoArgsConstructor`
to the class `OrderRequestBody` and `BookOrdered`.

TODO:

1. Since hibernate 6.2, `GenerationTime` is deprecated.
Therefore, the `Order` class should be modified.

2. The class `FileStorageService` is a little bit strange.
And its usage in `BookstoreApplication` may be not a good practice.
Try to re-implement the feature one day.